requirements.txt
dj-database-url (0.3.0)
Django (1.7.1)
django-pgjson (0.2.1)
djangorestframework (3.0.0)
gunicorn (19.1.1)
psycopg2 (2.5.4)
wsgiref (0.1.2)
